+++ .babelrc
@@ -10,7 +10,8 @@
 			"IE 11",
 			'last 2 Edge versions'
       	],
-        "useBuiltIns": "usage"
+        "useBuiltIns": "usage",
+        "corejs": 3
       }
     ]
   ]
