+++ dev/js/scripts/main.js
@@ -1,26 +1,16 @@
-import BrowserDetect from "./BrowserDetect";
-//import Responsive from "./Responsive";
-import SmoothScroll from "./SmoothScroll";
-//import Accordion from "./Accordion";
-//import Tab from "./Tab";
-import SpNavi from "./SpNavi";
-import ToTop from "./ToTop";
-//import DropDownMenu from "./DropDownMenu";
-//import BlockLink from "./BlockLink";
-//import Gmap from "./Gmap";
-//import matchheight from "jquery-match-height";
-import link_icon from "./link_icon";
-/* ===============================================
-Check responsive state
-=============================================== */
-// const r = new Responsive();
+import BrowserDetect from "modules/BrowserDetect";
+import SmoothScroll from "modules/SmoothScroll";
+import SpNavi from "modules/SpNavi";
+import ToTop from "modules/ToTop";
+import link_icon from "modules/link_icon";
+import "../../sass/style.scss";
 
-// $(window).on('resize',function(){
-// check is responsive;
-// 	console.log(r.state()());
-// });
-// check is mobile;
-// alert(r.isPhone());
+const req = require.context("../../pug/", false, /\.pug/);
+req.keys().forEach(fileName => {
+  const m = fileName.match(/^\.\/_/);
+  req(fileName);
+});
+
 /* ===============================================
 SP Navi
 =============================================== */
@@ -31,40 +21,8 @@ let param = {
   filter: "resposive_flg",
   speed: 200
 };
-
 const spnavi = new SpNavi(param);
 spnavi.exec();
-
-//execに可変引数で処理を渡してナビの開閉時に処理させるサンプル
-//例として、SPナビでアコーディオンを開くたびにiScrollをリセットする
-// let spnav_accordion_1 = (scroller)=>{
-// 	$('#scroller .inner > ul > li > span').off();
-// 	$('#scroller .inner > ul > li > span').on('click',function(){
-// 		let $parent = $(this).parent('li');
-// 		let $child = $(this).next('.gnavi-child');
-// 		let status = $child.css('display');
-
-// 		if(status == 'block'){
-// 			$child.slideUp(300,function(){
-// 				scroller.refresh();
-// 			});
-// 			$parent.removeClass('active');
-// 		}else{
-// 			$child.slideDown(300,function(){
-// 				scroller.refresh();
-// 			});
-// 			$parent.addClass('active');
-// 		}
-
-// 		return false;
-// 	});
-// }
-// spnavi.exec(spnav_accordion_1);
-/* ===============================================
-box link
-=============================================== */
-// let box_link = new BlockLink();
-// box_link.exec('.box_link');
 /* ===============================================
 Smooth Scroll
 =============================================== */
@@ -77,103 +35,6 @@ show/hide toggle
 const totop = new ToTop("#totop", "#global_footer", 100, 400);
 totop.exec();
 
-/**
- * 外部リンク・PDF・Word・Exlsに_blankの自動付与
- *
- */
-link_icon();
-/* ===============================================
-DropDown
-
-markp:
-<ul>
-	<li>
-		<a href=""></a>
-		<ul class="child">
-			<li>
-				<a href=""></a>
-			</li>
-			<li>
-				<a href=""></a>
-			</li>
-			<li>
-				<a href=""></a>
-			</li>
-		</ul>
-	</li>
-	<li>
-		<a href=""></a>
-	</li>
-	<li>
-		<a href=""></a>
-	</li>
-</ul>
-=============================================== */
-// const dd = new DropDownMenu('ul li','.child',200);
-// dd.exec();
-/* ===============================================
-Accordion
-
-markp:
-<dl>
-	<dt></dt>
-	<dd></dd>
-	<dt></dt>
-	<dd></dd>
-	<dt></dt>
-	<dd></dd>
-</dl>
-=============================================== */
-// const ac = new Accordion($('dt'));
-// ac.exec();
-/* ===============================================
-Tab
-
-markup:
-<div class="tc">
-	<ul>
-		<li>tab1</li>
-		<li>tab2</li>
-		<li>tab3</li>
-	</ul>
-	<div class="tc_content">
-		tabcon1tabcon1tabcon1tabcon1tabcon1tabcon1
-	</div>
-	<div class="tc_content">
-		tabcon2tabcon2tabcon2tabcon2tabcon2tabcon2
-	</div>
-	<div class="tc_content">
-		tabcon3tabcon3tabcon3tabcon3tabcon3tabcon3
-	</div>
-</div>
-=============================================== */
-// const tab = new Tab('.tc li','.tc .tc_content');
-// tab.exec();
-
-// /* gmap
-// ----------------------- */
-// let gmap = new Gmap();
-// $(window).on('load',function(){
-// 	gmap.exec();
-// });
-
-/* ===============================================
-instagram with infinityScroll
-
-markup:
-<div class="slide--instagram_wrapper" id="insragram">
-	<div class="slide--instagram">
-		<div id="js-slide--instagram">
-			<ul class="inner"></ul>
-		</div>
-	</div>
-</div>
-=============================================== */
-/**
- *
- * @param {sring} requet_url instagramのリクエストURL ./_instagram_template.htmlにhtmlテンプレートサンプルあり
- * @param {element} append_target 読み込んだ写真を追加する要素
- * @param {element} slide_target スライドショーさせる要素
- */
-// import instagram from './instagram';
-// instagram('request_url','#js-slide--instagram .inner','#js-slide--instagram');
+if (module.hot) {
+  module.hot.accept();
+}
