+++ dev/js/scripts/modules/BrowserDetect.js
