+++ dev/js/scripts/modules/FloatAnimation.js
