+++ dev/js/scripts/modules/SmoothScroll.js
