+++ dev/js/scripts/modules/Tab.js
