+++ dev/js/scripts/modules/infiniteSlide.js
