+++ dev/js/scripts/modules/jQuery.Pause.js
