+++ dev/js/scripts/modules/link_icon.js
