+++ dev/pug/_inc_scripts.pug
@@ -4,4 +4,4 @@
 //- else
 script(src='https://code.jquery.com/jquery-2.2.4.min.js', integrity='sha256-BbhdlvQf/xTY9gja0Dq3HiwQF8LaCRTXxZKRutelT44=', crossorigin='anonymous')
 script(src='https://cdnjs.cloudflare.com/ajax/libs/lodash.js/4.17.4/lodash.min.js')
-script(src=`/js/scripts/main.js`)
+script(src=`/js/bundle.js`)
