+++ dev/sass/helper/_shorthands.scss
@@ -162,28 +162,28 @@ a img.aligncenter {
   }
 }
 
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_mt#{$i * 1} {
     margin-top : #{$i * 1}px !important;
   }
 }
 
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_mb#{$i * 1} {
     margin-bottom : #{$i * 1}px !important;
   }
 }
 
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_ml#{$i * 1} {
     margin-left : #{$i * 1}px !important;
   }
 }
 
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_mr#{$i * 1} {
     margin-right : #{$i * 1}px !important;
@@ -202,7 +202,7 @@ a img.aligncenter {
     padding-left : #{$i * 5}px !important;
   }
 }
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_pl#{$i * 1} {
     padding-left : #{$i * 1}px !important;
@@ -221,7 +221,7 @@ a img.aligncenter {
   }
 }
 
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_pr#{$i * 1} {
     padding-right : #{$i * 1}px !important;
@@ -239,7 +239,7 @@ a img.aligncenter {
     padding-top : #{$i * 5}px !important;
   }
 }
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_pt#{$i * 1} {
     padding-top : #{$i * 1}px !important;
@@ -257,7 +257,7 @@ a img.aligncenter {
     padding-bottom : #{$i * 5}px !important;
   }
 }
-@for $i from -10 through 180 {
+@for $i from -10 through 70 {
 
   .large_pb#{$i * 1} {
     padding-bottom : #{$i * 1}px !important;
@@ -266,28 +266,28 @@ a img.aligncenter {
 
 @each $breakpoint, $value in $breakpoints {
   @media only screen and ( max-width : map-get( $breakpoints, $breakpoint ) ) {
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_mt#{$i * 1} {
         margin-top : #{$i * 1}px !important;
       }
     }
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_mb#{$i * 1} {
         margin-bottom : #{$i * 1}px !important;
       }
     }
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_ml#{$i * 1} {
         margin-left : #{$i * 1}px !important;
       }
     }
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_mr#{$i * 1} {
         margin-right : #{$i * 1}px !important;
@@ -297,7 +297,7 @@ a img.aligncenter {
 		------------------------------------- */
     // padding-left
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_pl#{$i * 1} {
         padding-left : #{$i * 1}px !important;
@@ -306,7 +306,7 @@ a img.aligncenter {
 
     // padding-right
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_pr#{$i * 1} {
         padding-right : #{$i * 1}px !important;
@@ -315,7 +315,7 @@ a img.aligncenter {
 
     // padding-top
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_pt#{$i * 1} {
         padding-top : #{$i * 1}px !important;
@@ -324,7 +324,7 @@ a img.aligncenter {
 
     // padding-bottom
 
-    @for $i from -10 through 180 {
+    @for $i from -10 through 70 {
 
       .#{$breakpoint}_pb#{$i * 1} {
         padding-bottom : #{$i * 1}px !important;
