+++ package.json
@@ -13,7 +13,7 @@
 		"browser-sync": "^2.18.7",
 		"css-loader": "^2.0.1",
 		"del": "^3.0.0",
-		"file-loader": "^2.0.0",
+		"file-loader": "^3.0.1",
 		"html-loader": "^0.5.5",
 		"lodash": "^4.17.4",
 		"path": "^0.12.7",
@@ -32,11 +32,13 @@
 		"bxslider": "^4.2.11",
 		"copy-webpack-plugin": "^5.0.2",
 		"cpx": "^1.5.0",
-		"cssnano": "^4.1.4",
+		"css-hot-loader": "^1.4.4",
+		"cssnano": "^4.1.10",
 		"eslint": "^5.16.0",
 		"eslint-config-prettier": "^4.1.0",
 		"eslint-plugin-prettier": "^3.0.1",
 		"evenizer": "^0.1.17",
+		"extract-loader": "^3.1.0",
 		"extract-text-webpack-plugin": "^4.0.0-beta.0",
 		"fs": "^0.0.1-security",
 		"globule": "^1.2.1",
@@ -51,6 +53,7 @@
 		"jquery-lineup": "^1.1.2",
 		"jquery-match-height": "^0.7.2",
 		"lint-staged": "^8.1.5",
+		"mini-css-extract-plugin": "^0.6.0",
 		"node-sass": "^4.5.3",
 		"node-sass-glob-importer": "^5.2.0",
 		"npm-run-all": "^4.0.2",
@@ -74,6 +77,7 @@
 		"uglifyjs-webpack-plugin": "^2.0.1",
 		"watch": "^1.0.2",
 		"webpack-dev-server": "^3.3.1",
+		"webpack-merge": "^4.2.1",
 		"webpack-notifier": "^1.5.0"
 	},
 	"scripts": {
@@ -81,8 +85,8 @@
 		"lint:js": "prettier-eslint --write ./dev/js/**/*.js",
 		"lint": "yarn lint:js & yarn lint:css",
 		"img-even": "evenizer -i ./dev/images/*.*",
-		"start": "export NODE_ENV=development & webpack-dev-server",
-		"build": "yarn lint:css & yarn lint:js & export NODE_ENV=production & webpack -p"
+		"start": "webpack-dev-server --hot --inline --config webpack.dev.js",
+		"build": "webpack --config webpack.production.js"
 	},
 	"husky": {
 		"hooks": {
@@ -94,9 +98,6 @@
 		"*.js": [
 			"prettier-eslint --write",
 			"git add"
-		],
-		"ignore": [
-			"**/build/**/*.js"
 		],
 		"*.scss": [
 			"prettier-stylelint --write",
