--- webpack.config.babel.js
@@ -1,197 +1,0 @@
-const path = require("path");
-const globule = require("globule");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-const webpack = require("webpack");
-const WebpackNotifierPlugin = require("webpack-notifier");
-const ExtractTextPlugin = require("extract-text-webpack-plugin");
-const globImporter = require("node-sass-glob-importer");
-const ImageminPlugin = require("imagemin-webpack-plugin").default;
-const ImageminJpegTran = require("imagemin-jpegtran");
-
-const supported = ["IE 11", "last 2 versions"];
-
-const MODE = process.env.NODE_ENV;
-
-const opts = {
-  srcDir: path.join(__dirname, "dev"),
-  destDir: path.join(__dirname, "build")
-};
-
-const convertExtensions = {
-  pug: "html",
-  scss: "css",
-  js: "js"
-};
-
-const files = {};
-Object.keys(convertExtensions).forEach(from => {
-  const to = convertExtensions[from];
-  if (from == "scss") {
-    files["css/style.css"] = "./dev/sass/style.scss";
-  } else if (from == "pug") {
-    globule
-      .find([`**/*.${from}`, `!**/_*.${from}`], { cwd: opts.srcDir })
-      .forEach(filename => {
-        let replaced = filename.replace(new RegExp(`.${from}$`, "i"), `.${to}`);
-        replaced = replaced.replace("pug/", "");
-        files[replaced] = path.join(opts.srcDir, filename);
-      });
-  } else {
-    files["js/scripts/main.js"] = "./dev/js/scripts/main.js";
-  }
-});
-
-module.exports = [
-  {
-    mode: MODE,
-    entry: files,
-    output: {
-      filename: "[name]",
-      path: opts.destDir
-    },
-    cache: true,
-    devtool: MODE !== "production" ? "inline-source-map" : "",
-    module: {
-      rules: [
-        {
-          test: /\.pug$/,
-          use: ExtractTextPlugin.extract({
-            use: [
-              {
-                loader: "html-loader"
-              },
-              {
-                loader: "pug-html-loader",
-                options: {
-                  pretty: true
-                }
-              }
-            ]
-          })
-        },
-        {
-          test: /\.scss$/,
-          use: ExtractTextPlugin.extract({
-            fallback: "style-loader",
-            use: [
-              {
-                loader: "css-loader",
-                options: {
-                  sourceMap: true,
-                  importLoaders: 1,
-                  url: false
-                }
-              },
-              {
-                loader: "postcss-loader",
-                options: {
-                  sourceMap: true,
-                  plugins: [
-                    require("autoprefixer")({
-                      browsers: supported
-                    }),
-                    require("cssnano"),
-                    require("postcss-assets")({
-                      loadPaths: ["dev/images/"],
-                      relative: "./dev/css"
-                    })
-                  ]
-                }
-              },
-              {
-                loader: "sass-loader",
-                options: {
-                  sourceMap: true,
-                  importer: globImporter()
-                }
-              }
-            ]
-          })
-        },
-        {
-          test: /\.js$/,
-          exclude: /(node_modules)/,
-          loader: "babel-loader"
-        },
-        { test: /\.html$/, loader: "html-loader" }
-      ]
-    },
-    externals: {
-      jquery: "jQuery"
-    },
-    plugins: [
-      new ExtractTextPlugin("[name]"),
-      new CopyWebpackPlugin([{ from: "./", to: "images/" }], {
-        context: "dev/images"
-      }),
-      new ImageminPlugin({
-        // GIF
-        test: /images\/([a-z_\-\s0-9]+)\.gif$/i,
-        gifsicle: {
-          interlaced: false,
-          optimizationLevel: 3,
-          colors: 256
-        }
-      }),
-
-      new ImageminPlugin({
-        // PNG
-        test: /images\/([a-z_\-\s0-9]+)\.png$/i,
-        optipng: {
-          optimizationLevel: 7,
-          bitDepthReduction: true,
-          colorTypeReduction: true,
-          paletteReduction: true
-        }
-      }),
-
-      new ImageminPlugin({
-        // JPG, JPEG
-        test: /images\/([a-z_\-\s0-9]+)\.jpe?g$/i,
-        jpegtran: {
-          progressive: true,
-          arithmetic: false
-        },
-        plugins: [
-          ImageminJpegTran({
-            quality: 85,
-            progressive: true,
-            targa: false,
-            revert: false,
-            dcScanOpt: 1,
-            notrellis: false,
-            notrellisDC: false,
-            tune: "hvs-psnr",
-            noovershoot: false,
-            arithmetic: false,
-            // quantTable : 0,
-            smooth: 0
-            // maxmemory  : Number,
-          })
-        ]
-      }),
-      new webpack.ProvidePlugin({
-        $: "jquery",
-        jQuery: "jquery",
-        "window.jQuery": "jquery"
-      }),
-      new WebpackNotifierPlugin({
-        title: "Success compiled!",
-        contentImage: path.join(__dirname, "dev/js/icons/shibasaki_ko.jpg"),
-        alwaysNotify: true
-      })
-    ],
-    performance: { hints: false },
-    devServer: {
-      contentBase: __dirname + "/",
-      watchContentBase: true,
-      open: true,
-      compress: true,
-      inline: true,
-      hot: true,
-      host: "0.0.0.0",
-      disableHostCheck: true,
-      useLocalIp: true
-    }
-  }
-];
